// packages/engine-core/src/types/input-event-type.ts
/**
 * Input event types
 */
export enum InputEventType {
    MOUSE_MOVE = 'mouse-move',
    MOUSE_DOWN = 'mouse-down',
    MOUSE_UP = 'mouse-up',
    MOUSE_ENTER = 'mouse-enter',
    MOUSE_LEAVE = 'mouse-leave',
    MOUSE_CLICK = 'mouse-click',
    KEY_DOWN = 'key-down',
    KEY_UP = 'key-up',
    WHEEL = 'wheel',
} 